chain: move Transparent{Input,Output} to transparent
Also bring the `Script` type there too.
