+++ zebra-chain/src/transparent/tests.rs
@@ -1,0 +1,1 @@
+mod arbitrary;
